+++ cdk/lib/db.ts
@@ -1,8 +1,8 @@
 import { Alarm, Metric, Unit, ComparisonOperator, TreatMissingData } from '@aws-cdk/aws-cloudwatch';
-import { Vpc, SecurityGroup, SubnetType, Peer, Port, InstanceType } from '@aws-cdk/aws-ec2';
+import { Vpc, SubnetType, Peer, InstanceType } from '@aws-cdk/aws-ec2';
 import { Secret } from '@aws-cdk/aws-secretsmanager';
 import { Secret as EcsSecret } from '@aws-cdk/aws-ecs';
-import { Construct, Stack, CfnOutput, SecretValue, Duration } from '@aws-cdk/core';
+import { Construct, Stack, CfnOutput, SecretValue, Duration, RemovalPolicy } from '@aws-cdk/core';
 import {
   DatabaseCluster as DocDbDatabaseCluster,
   ClusterParameterGroup as DocDbClusterParameterGroup,
@@ -19,6 +19,7 @@ const DEFAULT_DB_INSTANCE_TYPE = 't3.medium';
 const DEFAULT_AURORA_MYSQL_ENGINE_VERSION = '5.7.mysql_aurora.2.04.9'; // current LTS
 const DEFAULT_DOCDB_ENGINE_VERSION = '3.6';
 const DEFAULT_DOCDB_PARAM_GROUP_FAMILY = 'docdb3.6';
+const DEFAULT_REMOVAL_POLICY = 'DESTROY';
 
 export interface CacclDbOptions {
   // currently either 'docdb' or 'mysql'
@@ -35,6 +36,8 @@ export interface CacclDbOptions {
   profiler?: boolean,
   // only used by mysql; provisioning will create the named database
   databaseName?: string,
+  // removal policy controls what happens to the db if it's replaced or otherwise stops being managed by CloudFormation
+  removalPolicy?: string,
 }
 
 export interface CacclDbProps {
@@ -66,6 +69,10 @@ export class CacclDb extends Construct {
 
   getDashboardLink: Function;
 
+  removalPolicy: RemovalPolicy;
+
+  pgRemovalPolicy: RemovalPolicy;
+
   static createDbConstruct(scope: Construct, props: CacclDbProps) {
     const { options } = props;
     switch (options.engine.toLowerCase()) {
@@ -80,6 +87,18 @@ export class CacclDb extends Construct {
 
   constructor(scope: Construct, id: string, props: CacclDbProps) {
     super(scope, id);
+
+    const {
+      removalPolicy = DEFAULT_REMOVAL_POLICY,
+    } = props.options;
+
+    // removal policy for the cluster & instances
+    this.removalPolicy = (<any>RemovalPolicy)[removalPolicy];
+    // if we're keeping the dbs we have to keep the parameter group
+    // or cloudformation will barf
+    this.pgRemovalPolicy = this.removalPolicy === RemovalPolicy.RETAIN
+      ? RemovalPolicy.RETAIN
+      : RemovalPolicy.DESTROY;
 
     this.dbPasswordSecret = new Secret(this, 'DbPasswordSecret', {
       description: `docdb master user password for ${Stack.of(this).stackName}`,
@@ -197,6 +216,7 @@ export class CacclDocDb extends CacclDb {
       profiler = false,
     } = props.options;
 
+
     if (profiler) {
       this.clusterParameterGroupParams.profiler = 'enabled';
       this.clusterParameterGroupParams.profiler_threshold_ms = '500';
@@ -225,15 +245,14 @@ export class CacclDocDb extends CacclDb {
       backup: {
         retention: Duration.days(14),
       },
+      removalPolicy: this.removalPolicy,
     });
+
+    // this needs to happen after the parameter group has been associated with the cluster
+    parameterGroup.applyRemovalPolicy(this.pgRemovalPolicy);
 
     this.host = this.dbCluster.clusterEndpoint.hostname;
     this.port = this.dbCluster.clusterEndpoint.portAsString();
-
-    // add an ingress rule to the db security group
-    // const dbSg = SecurityGroup.fromSecurityGroupId(this, 'DocDbSecurityGroup', this.dbCluster.securityGroupId);
-    // const ingressPort = Port.tcp(+this.port)
-    // dbSg.addIngressRule(Peer.ipv4(vpc.vpcCidrBlock), ingressPort);
 
     appEnv.addEnvironmentVar('MONGO_USER', 'root');
     appEnv.addEnvironmentVar('MONGO_HOST', `${this.host}:${this.port}`);
@@ -313,8 +332,13 @@ export class CacclRdsDb extends CacclDb {
       },
       backup: {
         retention: Duration.days(14),
-      }
+      },
+      removalPolicy: this.removalPolicy,
     });
+
+    // this needs to happen after the parameter group has been associated with the cluster
+    clusterParameterGroup.applyRemovalPolicy(this.pgRemovalPolicy);
+    instanceParameterGroup.applyRemovalPolicy(this.pgRemovalPolicy);
 
     // for rds/mysql we do NOT include the port as part of the host value
     this.host = this.dbCluster.clusterEndpoint.hostname;